requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
asgiref==3.9.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
boto3==1.39.9
botocore==1.39.9
certifi==2025.7.14
cffi==1.17.1
channels==4.3.0
# channels[daphne]
channels_redis
charset-normalizer==3.4.2
constantly==23.10.4
cryptography==45.0.5
daphne==4.2.1
Django==5.2.4
django-cors-headers==4.5.0
django-storages==1.14.6
djangorestframework==3.15.1
edge-tts==7.0.2
frozenlist==1.7.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
jmespath==1.0.1
multidict==6.6.3
mysql-connector-python==9.3.0
mysqlclient==2.2.7
pillow==11.3.0
propcache==0.3.2
psycopg2==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydub==0.25.1
pyOpenSSL==25.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python3-pika==0.9.14
PyYAML==6.0.2
requests==2.32.4
s3transfer==0.13.1
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
srt==3.5.3
tabulate==0.9.0
Twisted==25.5.0
txaio==25.6.1
typing_extensions==4.14.1
urllib3==2.5.0
yarl==1.20.1
zope.interface==7.2
